# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Club Connect - Brooklyn College Computer Science Club
  description: The OpenAPI document for Club Connect, the full-stack portal for the Brooklyn College Computer Science Club.
  version: 1.0.0
  contact:
    email: contact@bccompsci.club
servers:
  - url: 'https://api.bccompsci.club'
  - url: 'http://127.0.0.1:8080'
...
components:
  securitySchemes:
    jwtCookieAuth:
      type: apiKey
      in: cookie
      name: token
paths:
...
  /events:
    get:
      summary: Returns a list of events.
      tags:
        - events
      responses:
        200:
          description: A JSON array of events.
          content:
            application/json:
              schema:
                type: array
    post:
      summary: Creates a new event. Not the same as creating an event request.
      description: Requires the Admin role.
      tags:
        - events
      security:
        - jwtCookieAuth: []
      responses:
        201:
          description: The event has been created.
          headers:
            Location:
              description: The location of the new event.
              schema:
                type: string
  /events/request:
    post:
      summary: Submits a new event request. Not the same as creating an event.
      tags:
        - events
      security:
        - jwtCookieAuth: []
      responses:
        200:
          description: The event request has been created.
  /events/{id}:
    get:
      summary: Gets an event by its ID.
      tags:
        - events
      parameters:
        - name: id
          in: path
          description: Event ID
          required: true
          schema:
            type: integer
      responses:
        200:
          description: A JSON object containing event data.
          content:
            application/json:
              schema:
                type: object
        404:
          description: No event was found with the ID.
    delete:
      summary: Deletes an event by its ID.
      description: Requires the Admin role.
      tags:
        - events
      security:
        - jwtCookieAuth: []
      responses:
        200:
          description: The event has been deleted.